# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatUploadFinalizeResult": {
        "additionalProperties": false,
        "properties": {
          "bound": {
            "const": false,
            "type": "boolean"
          },
          "scope": {
            "const": "chat",
            "type": "string"
          },
          "status": {
            "const": "confirmed",
            "type": "string"
          },
          "upload": {
            "$ref": "#/components/schemas/ConfirmedUpload"
          }
        },
        "required": [
          "scope",
          "status",
          "bound",
          "upload"
        ],
        "type": "object"
      },
...
      "ConfirmedUpload": {
        "additionalProperties": false,
        "properties": {
          "byte_size": {
            "minimum": 1,
            "type": "integer"
          },
          "confirmed_at": {
            "format": "date-time",
            "type": "string"
          },
          "content_type": {
            "minLength": 1,
            "type": "string"
          },
          "duration": {
            "minimum": 1,
            "type": [
              "integer",
              "null"
            ]
          },
          "filename": {
            "maxLength": 255,
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "object_key": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string"
          },
          "poster_upload_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "scope": {
            "$ref": "#/components/schemas/MediaScope"
          },
          "target_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope",
          "target_id",
          "object_key",
          "kind",
          "content_type",
          "byte_size",
          "duration",
          "filename",
          "confirmed_at",
          "poster_upload_id"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "description": "Reserved for a future additive contract; exactly null in v1",
            "type": "null"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "MediaAccessUrl": {
        "additionalProperties": false,
        "properties": {
          "byte_size": {
            "minimum": 1,
            "type": "integer"
          },
          "content_type": {
            "minLength": 1,
            "type": "string"
          },
          "duration": {
            "minimum": 1,
            "type": [
              "integer",
              "null"
            ]
          },
          "expires_in": {
            "const": 600,
            "type": "integer"
          },
          "filename": {
            "maxLength": 255,
            "type": [
              "string",
              "null"
            ]
          },
          "height": {
            "minimum": 1,
            "type": [
              "integer",
              "null"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "media_upload_id": {
            "format": "uuid",
            "type": "string"
          },
          "url": {
            "format": "uri",
            "type": "string"
          },
          "width": {
            "minimum": 1,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "media_upload_id",
          "url",
          "content_type",
          "byte_size",
          "width",
          "height",
          "duration",
          "filename",
          "expires_in"
        ],
        "type": "object"
      },
      "MediaKind": {
        "enum": [
          "image",
          "video",
          "audio"
        ],
        "type": "string"
      },
...
      "MediaScope": {
        "enum": [
          "chat"
        ],
        "type": "string"
      },
...
      "PresignedPut": {
        "additionalProperties": false,
        "properties": {
          "expires_in": {
            "const": 3600,
            "type": "integer"
          },
          "url": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_in"
        ],
        "type": "object"
      },
...
      "UploadFinalize": {
        "additionalProperties": false,
        "properties": {
          "height": {
            "minimum": 1,
            "type": [
              "integer",
              "null"
            ]
          },
          "poster_upload_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "width": {
            "minimum": 1,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "UploadFinalizeResult": {
        "$ref": "#/components/schemas/ChatUploadFinalizeResult"
      },
      "UploadIntent": {
        "additionalProperties": false,
        "properties": {
          "byte_size": {
            "minimum": 1,
            "type": "integer"
          },
          "content_type": {
            "minLength": 1,
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "filename": {
            "maxLength": 255,
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "object_key": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/MediaScope"
          },
          "target_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope",
          "target_id",
          "object_key",
          "kind",
          "content_type",
          "byte_size",
          "filename",
          "expires_at",
          "created_at"
        ],
        "type": "object"
      },
      "UploadIntentCreate": {
        "additionalProperties": false,
        "allOf": [
          {
            "if": {
              "properties": {
                "content_type": {
                  "enum": [
                    "image/jpeg",
                    "image/png",
                    "image/webp",
                    "image/gif"
                  ]
                }
              },
              "required": [
                "content_type"
              ]
            },
            "then": {
              "properties": {
                "byte_size": {
                  "maximum": 10485760
                }
              }
            }
          },
          {
            "if": {
              "properties": {
                "content_type": {
                  "const": "video/mp4"
                }
              },
              "required": [
                "content_type"
              ]
            },
            "then": {
              "properties": {
                "byte_size": {
                  "maximum": 52428800
                }
              }
            }
          },
          {
            "if": {
              "properties": {
                "content_type": {
                  "enum": [
                    "audio/webm",
                    "audio/mp4",
                    "audio/ogg"
                  ]
                }
              },
              "required": [
                "content_type"
              ]
            },
            "then": {
              "properties": {
                "byte_size": {
                  "maximum": 15728640
                }
              }
            }
          }
        ],
        "properties": {
          "byte_size": {
            "maximum": 52428800,
            "minimum": 1,
            "type": "integer"
          },
          "content_type": {
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp",
              "image/gif",
              "video/mp4",
              "audio/webm",
              "audio/mp4",
              "audio/ogg"
            ],
            "type": "string"
          },
          "filename": {
            "maxLength": 255,
            "type": [
              "string",
              "null"
            ]
          },
          "scope": {
            "$ref": "#/components/schemas/MediaScope"
          },
          "target_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "target_id",
          "content_type",
          "byte_size"
        ],
        "type": "object"
      },
      "UploadIntentWithPresignedPut": {
        "additionalProperties": false,
        "properties": {
          "put": {
            "$ref": "#/components/schemas/PresignedPut"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadIntent"
          }
        },
        "required": [
          "upload",
          "put"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "bearerFormat": "JWT",
        "description": "Short-lived signed access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Production HTTP API for Jamye mobile clients. All error responses use the shared ErrorEnvelope schema.",
    "license": {
      "name": ""
    },
    "title": "Jamye Server API",
    "version": "1"
  },
...
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/media/uploads": {
      "post": {
        "operationId": "MD1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadIntentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadIntentWithPresignedPut"
                }
              }
            },
            "description": "Resource created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response; inspect error.code for the stable machine-readable reason"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Create a media upload intent",
        "tags": [
          "media"
        ],
        "x-jamye-authorization": "member+target policy",
        "x-jamye-behavior-test": "tests/media/upload_finalize_http.rs::md1_returns_a_server_minted_intent_and_constrained_put",
        "x-jamye-errors": "E5",
        "x-jamye-fixture": "contracts/contributions/task-8/fixtures/media-flow.json",
        "x-jamye-owner-contribution": "contracts/contributions/task-8/dto/operations.json"
      }
    },
    "/api/v1/media/uploads/{upload_id}/finalize": {
      "post": {
        "operationId": "MD2",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadFinalize"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadFinalizeResult"
                }
              }
            },
            "description": "Successful response or canonical idempotent retry"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response; inspect error.code for the stable machine-readable reason"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Finalize a media upload",
        "tags": [
          "media"
        ],
        "x-jamye-authorization": "upload owner+target policy",
        "x-jamye-behavior-test": "tests/media/upload_finalize_http.rs::md2_chat_finalize_returns_the_confirmed_unbound_capability",
        "x-jamye-errors": "E5",
        "x-jamye-fixture": "contracts/contributions/task-8/fixtures/media-flow.json",
        "x-jamye-owner-contribution": "contracts/contributions/task-8/dto/operations.json"
      }
    },
    "/api/v1/media/{media_id}/download": {
      "get": {
        "operationId": "MD5",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Temporary redirect to a short-lived signed URL",
            "headers": {
              "Location": {
                "description": "Short-lived authorized download URL",
                "schema": {
                  "format": "uri",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response; inspect error.code for the stable machine-readable reason"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Redirect to a media download URL",
        "tags": [
          "media"
        ],
        "x-jamye-authorization": "member",
        "x-jamye-behavior-test": "tests/media/access_http.rs::md5_returns_an_empty_307_redirect_to_the_authorized_download_url",
        "x-jamye-errors": "E5",
        "x-jamye-fixture": "contracts/contributions/task-8/fixtures/media-flow.json",
        "x-jamye-owner-contribution": "contracts/contributions/task-8/dto/operations.json"
      }
    },
    "/api/v1/media/{media_id}/url": {
      "get": {
        "operationId": "MD4",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAccessUrl"
                }
              }
            },
            "description": "Successful response or canonical idempotent retry"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response; inspect error.code for the stable machine-readable reason"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Issue a short-lived media view URL",
        "tags": [
          "media"
        ],
        "x-jamye-authorization": "member",
        "x-jamye-behavior-test": "tests/media/access_http.rs::md4_returns_only_public_metadata_and_the_short_reissued_url",
        "x-jamye-errors": "E5",
        "x-jamye-fixture": "contracts/contributions/task-8/fixtures/media-flow.json",
        "x-jamye-owner-contribution": "contracts/contributions/task-8/dto/operations.json"
      }
    },
...
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://jamye-api.ridewithmin.com"
    }
  ],
...
}